refactor(integration-tests): fix Renovate tracking for Apicurio Registry image
## Problem

The Apicurio Registry container image reference in RecordValidationBaseIT
had become stale (v3.1.6) while JsonSchemaRecordValidationTlsIT was being
updated to v3.3.0 by Renovate. This inconsistency occurred because:

1. RecordValidationBaseIT.java was not included in the Renovate config
2. JsonSchemaRecordValidationTlsIT duplicated the image reference
3. Renovate only tracked JsonSchemaRecordValidation*IT.java files

This meant most schema validation tests (Avro, Protobuf, and non-TLS JSON
Schema) were using an outdated registry version, missing security patches
and bug fixes.

## Solution

Refactored the test class hierarchy to eliminate duplication and ensure
Renovate tracks the single source of truth:

1. Created RecordSchemaValidationBaseIT - intermediate base class for tests
   requiring Apicurio Registry (Avro, JSON Schema, Protobuf validation)

2. Moved Apicurio-specific infrastructure to RecordSchemaValidationBaseIT:
   - APICURIO_REGISTRY_IMAGE constant (single source of truth)
   - apicurioRegistryDockerImageName() method (extracted common logic)
   - startRegistryContainer(), registryUrl(), registryClient() helpers

3. Cleaned up RecordValidationBaseIT to only contain generic test utilities
   (used by all validation tests, not just schema-based ones)

4. Updated test hierarchy:
   - AvroRecordValidationIT → extends RecordSchemaValidationBaseIT
   - JsonSchemaRecordValidationIT → extends RecordSchemaValidationBaseIT
   - JsonSchemaRecordValidationTlsIT → extends RecordSchemaValidationBaseIT
   - ProtobufRecordValidationIT → extends RecordSchemaValidationBaseIT
   - JsonSyntaxRecordValidationIT → extends RecordValidationBaseIT (no registry)
   - JwsSignatureRecordValidationIT → extends RecordValidationBaseIT (no registry)

5. Updated Renovate config to track RecordSchemaValidationBaseIT.java

## Benefits

- Single source of truth for Apicurio Registry image reference
- All schema validation tests now use the same registry version
- Renovate will update all schema tests automatically
- Clearer class hierarchy reflecting actual dependencies
- No duplication of DockerImageName parsing logic
