+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/it/filter/validation/RecordValidationBaseIT.java
@@ -15,13 +15,6 @@ import java.util.concurrent.Future;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.testcontainers.DockerClientFactory;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.containers.wait.strategy.Wait;
-import org.testcontainers.utility.DockerImageName;
-
-import io.apicurio.registry.client.RegistryClientFactory;
-import io.apicurio.registry.client.common.RegistryClientOptions;
-import io.apicurio.registry.rest.client.RegistryClient;
 
 import io.kroxylicious.it.BaseIT;
 import io.kroxylicious.testing.integration.tester.KroxyliciousTester;
@@ -32,10 +25,6 @@ import static org.apache.kafka.clients.consumer.ConsumerConfig.GROUP_ID_CONFIG;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public abstract class RecordValidationBaseIT extends BaseIT {
-
-    private static final String APICURIO_REGISTRY_IMAGE = "quay.io/apicurio/apicurio-registry:3.1.6@sha256:d0625211cebb1f58a2982df29cb0945249d8f88f37ccce9e162c0c12c2aea89e";
-    private static final String APICURIO_REGISTRY_API = "/apis/registry/v3";
-    private static final int CONTAINER_PORT = 8080;
 
     public void assertThatFutureSucceeds(Future<RecordMetadata> future) {
         assertThat(future)
@@ -55,33 +44,6 @@ public abstract class RecordValidationBaseIT extends BaseIT {
         try (var consumer = tester.consumer(Map.of(GROUP_ID_CONFIG, UUID.randomUUID().toString(), AUTO_OFFSET_RESET_CONFIG, "earliest"))) {
             consumer.subscribe(Set.of(topic.name()));
             return consumer.poll(Duration.ofSeconds(10));
-        }
-    }
-
-    protected static GenericContainer<?> startRegistryContainer() {
-        String image = APICURIO_REGISTRY_IMAGE;
-        DockerImageName dockerImageName = DockerImageName.parse(image)
-                .asCompatibleSubstituteFor(DockerImageName.parse(image.substring(0, image.indexOf("@"))));
-
-        GenericContainer<?> container = new GenericContainer<>(dockerImageName)
-                .withExposedPorts(CONTAINER_PORT)
-                .waitingFor(Wait.forHttp(APICURIO_REGISTRY_API + "/system/info").forStatusCode(200));
-
-        container.start();
-        return container;
-    }
-
-    protected static String registryUrl(GenericContainer<?> container) {
-        return "http://" + container.getHost() + ":" + container.getMappedPort(CONTAINER_PORT) + APICURIO_REGISTRY_API;
-    }
-
-    protected static RegistryClient registryClient(String registryUrl) {
-        return RegistryClientFactory.create(RegistryClientOptions.create(registryUrl));
-    }
-
-    protected static void stopContainer(GenericContainer<?> container) {
-        if (container != null && container.isRunning()) {
-            container.stop();
         }
     }
